gemu/requirements.txt
pyyaml
yara-python